+++ VulkanMemoryAllocator/changelog.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## WIP
+- Add Vulkan 1.3 functions to VulkanFunctions
 
 ## [0.10.2] - 2022-09-24
 - Raise upper bound on `vulkan`